use dynamic imports on image viewer
